[RENAME] Drop the repetitive prefix Spark in front
of all the Spark handlers. It does make sense to keep it in front
of `SparkDDF` and `SparkDDFManager`, since these are user-facing
and we want the user to know what kind of DDFs they are dealing with.

For the others, the naming is just too long, and there is little chance
of conflict/confusion with the base API.